backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
pydantic==2.11.7
python-multipart==0.0.20
pytest==8.4.1
httpx==0.28.1
numpy==2.1.3
scikit-learn==1.6.1
Pillow==11.3.0
